Moved UserOrdersPage.tsx to main pages folder, removed some unnecessary imports, updated router to accomodate the page location change
